refactor: split main/index.ts monolith into IPC handler modules
Extract 124 IPC handlers from single 3046-line file into 10 domain
modules in apps/desktop/src/main/handlers/:

| Module              | Lines | Handlers                              |
|---------------------|-------|---------------------------------------|
| noteHandlers.ts     |   429 | notes:*, tags:*, links:*, embeds:*    |
| notebookHandlers.ts |   311 | notebooks:*                           |
| authSyncHandlers.ts |   583 | auth:*, sync:*, encryption:*, etc.    |
| pluginHandlers.ts   |   410 | pluginConfig:*, plugins:*             |
| dataHandlers.ts     |   207 | data:* (backup, export, import)       |
| gitHandlers.ts      |   154 | git:*                                 |
| updateHandlers.ts   |   132 | updates:* + auto-updater init         |
| logHandlers.ts      |    53 | log:*                                 |
| aiKeyHandlers.ts    |    36 | ai:saveKey/getKey/removeKey/hasKey     |
| types.ts            |    62 | Shared dependency interfaces          |

index.ts reduced from 3046 → 887 lines (bootstrap + window management).
