+++ apps/desktop/src/main/index.ts
@@ -2,6 +2,7 @@
  * Electron Main Process
  *
  * Initializes the app, database, and IPC handlers.
+ * Handler logic is in apps/desktop/src/main/handlers/*.ts
  */
 
 // Initialize Sentry FIRST (before any other imports that might throw)
@@ -9,76 +10,26 @@
 import { initSentry } from './sentry';
 initSentry();
 
-import { join, normalize, basename } from 'path';
-import { readFile, writeFile, unlink, mkdir, rm, readdir, stat, rename } from 'fs/promises';
-import { existsSync, readFileSync, writeFileSync } from 'fs';
-import { execFile } from 'child_process';
-import { app, BrowserWindow, ipcMain, dialog, shell, protocol, net, nativeTheme } from 'electron';
-import { autoUpdater } from 'electron-updater';
-// electron-devtools-installer is imported dynamically below (dev only)
-
-/** Safely send IPC message to all windows, ignoring destroyed ones */
-function broadcastToWindows(channel: string, ...args: unknown[]): void {
-  for (const win of BrowserWindow.getAllWindows()) {
-    try {
-      if (!win.isDestroyed() && !win.webContents.isDestroyed()) {
-        win.webContents.send(channel, ...args);
-      }
-    } catch {
-      // Window was destroyed between check and send — ignore
-    }
-  }
-}
-import {
-  runMigrations,
-  createDataPaths,
-  createBackup,
-  listBackups,
-  restoreBackup,
-  exportNotes,
-  importNotes,
-  detectImportType,
-  type DataPaths,
-} from '@readied/storage-core';
+import { join, normalize } from 'path';
+import { readFile, writeFile, unlink } from 'fs/promises';
+import { readFileSync, writeFileSync, existsSync } from 'fs';
+import { app, BrowserWindow, ipcMain, protocol, net, nativeTheme } from 'electron';
+import { runMigrations, createDataPaths, type DataPaths } from '@readied/storage-core';
 import {
   createDatabase,
   allMigrations,
   SQLiteNoteRepository,
   SQLiteNotebookRepository,
 } from '@readied/storage-sqlite';
-import {
-  createNoteOperation,
-  updateNoteOperation,
-  updateTitleOperation,
-  deleteNoteOperation,
-  getNoteOperation,
-  archiveNoteOperation,
-  restoreNoteOperation,
-  duplicateNoteOperation,
-  moveNoteToNotebook,
-  pinNote,
-  unpinNote,
-  softDeleteNote,
-  restoreDeletedNote,
-  setNoteStatus,
-  type NoteStatus,
-} from '@readied/core';
-import {
-  createNoteId,
-  createNotebookId,
-  createNotebook,
-  renameNotebook,
-  moveNotebook,
-  createTag,
-  INBOX_NOTEBOOK_ID,
-} from '@readied/core';
-import {
-  type LicenseStorage,
-  type StoredTrialData,
-  type StoredLicenseData,
-  type StoredSubscriptionData,
+import { createNoteId, createNoteOperation, type NoteStatus } from '@readied/core';
+// eslint-disable-next-line @typescript-eslint/no-deprecated
+import type {
+  LicenseStorage,
+  StoredTrialData,
+  StoredLicenseData,
+  StoredSubscriptionData,
 } from '@readied/licensing';
-import { initLogger, createChildLogger, loggers, getLogger, type LogLevel } from './logger';
+import { initLogger, getLogger, loggers } from './logger';
 import { TokenStorage } from './services/tokenStorage.js';
 import { AiKeyStorage } from './services/aiKeyStorage.js';
 import { getOrCreateDeviceInfo, type DeviceInfo } from './services/deviceInfo.js';
@@ -88,13 +39,24 @@ import { SyncService } from './services/syncService.js';
 import { GitService } from './services/gitService.js';
 import { registerLicenseHandlers } from './handlers/licenseHandlers.js';
 import { registerShareHandlers } from './handlers/shareHandlers.js';
-import { scanPlugins } from './pluginScanner.js';
+import { registerNoteHandlers } from './handlers/noteHandlers.js';
+import { registerNotebookHandlers } from './handlers/notebookHandlers.js';
+import { registerDataHandlers } from './handlers/dataHandlers.js';
+import { registerLogHandlers } from './handlers/logHandlers.js';
+import { registerUpdateHandlers, initAutoUpdater } from './handlers/updateHandlers.js';
+import { registerAuthSyncHandlers } from './handlers/authSyncHandlers.js';
+import { registerGitHandlers } from './handlers/gitHandlers.js';
+import { registerPluginHandlers } from './handlers/pluginHandlers.js';
+import { registerAiKeyHandlers } from './handlers/aiKeyHandlers.js';
 import { startPluginWatcher, stopPluginWatcher } from './pluginWatcher.js';
 import { createAIService, getToolRegistry } from './ai/setup.js';
 import { registerBuiltInTools } from './ai/built-in-tools.js';
 import { registerAIHandlers as registerAIHandlersNew } from './ai/ipc-ai.js';
 
-// Database and repository (initialized on app ready)
+// ============================================================================
+// Global State
+// ============================================================================
+
 let db: ReturnType<typeof createDatabase> | null = null;
 let noteRepository: SQLiteNoteRepository | null = null;
 let notebookRepository: SQLiteNotebookRepository | null = null;
@@ -114,7 +76,61 @@ let pendingAuthToken: string | null = null;
 // Git service (initialized on app ready)
 let gitService: GitService | null = null;
 
-/** File-based license storage implementation */
+// ============================================================================
+// Shared Helpers (exported for handler modules)
+// ============================================================================
+
+/** Safely send IPC message to all windows, ignoring destroyed ones */
+export function broadcastToWindows(channel: string, ...args: unknown[]): void {
+  for (const win of BrowserWindow.getAllWindows()) {
+    try {
+      if (!win.isDestroyed() && !win.webContents.isDestroyed()) {
+        win.webContents.send(channel, ...args);
+      }
+    } catch {
+      // Window was destroyed between check and send — ignore
+    }
+  }
+}
+
+/** Helper to convert a Note to a snapshot for IPC */
+export function noteToSnapshot(note: {
+  id: string;
+  notebookId: string;
+  content: string;
+  title: string;
+  isPinned: boolean;
+  isDeleted: boolean;
+  status: NoteStatus;
+  metadata: {
+    createdAt: string;
+    updatedAt: string;
+    tags: readonly string[];
+    wordCount: number;
+    archivedAt: string | null;
+  };
+}) {
+  return {
+    id: note.id,
+    notebookId: note.notebookId,
+    content: note.content,
+    title: note.title,
+    createdAt: note.metadata.createdAt,
+    updatedAt: note.metadata.updatedAt,
+    tags: [...note.metadata.tags],
+    wordCount: note.metadata.wordCount,
+    archivedAt: note.metadata.archivedAt,
+    isArchived: note.metadata.archivedAt !== null,
+    isPinned: note.isPinned,
+    isDeleted: note.isDeleted,
+    status: note.status,
+  };
+}
+
+// ============================================================================
+// File-based License Storage
+// ============================================================================
+
 class FileLicenseStorage implements LicenseStorage {
   private licensePath: string;
   private trialPath: string;
@@ -225,6 +241,10 @@ function saveWindowState(state: WindowState): void {
   }
 }
 
+// ============================================================================
+// Initialization
+// ============================================================================
+
 /** Initialize data paths */
 function initDataPaths(): DataPaths {
   const userDataPath = app.getPath('userData');
@@ -250,6 +270,10 @@ function initDatabase(): void {
 
   dbLog.info('Database initialized');
 }
+
+// ============================================================================
+// Window Creation
+// ============================================================================
 
 /** Create the main window */
 function createWindow(): void {
@@ -426,286 +450,11 @@ function createSettingsWindow(): void {
   }
 }
 
-/** Helper to convert a Note to a snapshot for IPC */
-function noteToSnapshot(note: {
-  id: string;
-  notebookId: string;
-  content: string;
-  title: string;
-  isPinned: boolean;
-  isDeleted: boolean;
-  status: NoteStatus;
-  metadata: {
-    createdAt: string;
-    updatedAt: string;
-    tags: readonly string[];
-    wordCount: number;
-    archivedAt: string | null;
-  };
-}) {
-  return {
-    id: note.id,
-    notebookId: note.notebookId,
-    content: note.content,
-    title: note.title,
-    createdAt: note.metadata.createdAt,
-    updatedAt: note.metadata.updatedAt,
-    tags: [...note.metadata.tags],
-    wordCount: note.metadata.wordCount,
-    archivedAt: note.metadata.archivedAt,
-    isArchived: note.metadata.archivedAt !== null,
-    isPinned: note.isPinned,
-    isDeleted: note.isDeleted,
-    status: note.status,
-  };
-}
+// ============================================================================
+// Window Management IPC (small enough to keep inline)
+// ============================================================================
 
-/** Register IPC handlers for notes CRUD */
-function registerIpcHandlers(): void {
-  if (!noteRepository) {
-    throw new Error('Note repository not initialized');
-  }
-
-  const repo = noteRepository;
-
-  // Create note
-  ipcMain.handle(
-    'notes:create',
-    async (_event, input: { content: string; id?: string; notebookId?: string }) => {
-      return createNoteOperation(input, repo);
-    }
-  );
-
-  // Get note
-  ipcMain.handle('notes:get', async (_event, id: string) => {
-    const noteId = createNoteId(id);
-    return getNoteOperation({ id: noteId }, repo);
-  });
-
-  // Update note content
-  ipcMain.handle('notes:update', async (_event, input: { id: string; content: string }) => {
-    const noteId = createNoteId(input.id);
-    return updateNoteOperation({ id: noteId, content: input.content }, repo);
-  });
-
-  // Update note title (structural, independent from content)
-  ipcMain.handle('notes:updateTitle', async (_event, input: { id: string; title: string }) => {
-    const noteId = createNoteId(input.id);
-    return updateTitleOperation({ id: noteId, title: input.title }, repo);
-  });
-
-  // Delete note
-  ipcMain.handle('notes:delete', async (_event, id: string) => {
-    const noteId = createNoteId(id);
-    return deleteNoteOperation({ id: noteId }, repo);
-  });
-
-  // Archive note
-  ipcMain.handle('notes:archive', async (_event, id: string) => {
-    const noteId = createNoteId(id);
-    return archiveNoteOperation({ id: noteId }, repo);
-  });
-
-  // Restore note
-  ipcMain.handle('notes:restore', async (_event, id: string) => {
-    const noteId = createNoteId(id);
-    return restoreNoteOperation({ id: noteId }, repo);
-  });
-
-  // Duplicate note
-  ipcMain.handle('notes:duplicate', async (_event, id: string) => {
-    const noteId = createNoteId(id);
-    return duplicateNoteOperation({ id: noteId }, repo);
-  });
-
-  // Move note to notebook
-  ipcMain.handle('notes:move', async (_event, noteId: string, notebookId: string) => {
-    const note = await repo.get(createNoteId(noteId));
-    if (!note) {
-      return { ok: false, error: { type: 'NOT_FOUND', id: noteId } };
-    }
-
-    const movedNote = moveNoteToNotebook(note, createNotebookId(notebookId));
-    await repo.save(movedNote);
-
-    return {
-      ok: true,
-      data: noteToSnapshot(movedNote),
-    };
-  });
-
-  // Pin note
-  ipcMain.handle('notes:pin', async (_event, id: string) => {
-    const note = await repo.get(createNoteId(id));
-    if (!note) {
-      return { ok: false, error: { type: 'NOT_FOUND', id } };
-    }
-
-    const pinnedNote = pinNote(note);
-    await repo.save(pinnedNote);
-
-    return { ok: true, data: noteToSnapshot(pinnedNote) };
-  });
-
-  // Unpin note
-  ipcMain.handle('notes:unpin', async (_event, id: string) => {
-    const note = await repo.get(createNoteId(id));
-    if (!note) {
-      return { ok: false, error: { type: 'NOT_FOUND', id } };
-    }
-
-    const unpinnedNote = unpinNote(note);
-    await repo.save(unpinnedNote);
-
-    return { ok: true, data: noteToSnapshot(unpinnedNote) };
-  });
-
-  // Soft delete (move to trash)
-  ipcMain.handle('notes:softDelete', async (_event, id: string) => {
-    const note = await repo.get(createNoteId(id));
-    if (!note) {
-      return { ok: false, error: { type: 'NOT_FOUND', id } };
-    }
-
-    const deletedNote = softDeleteNote(note);
-    await repo.save(deletedNote);
-
-    return { ok: true, data: noteToSnapshot(deletedNote) };
-  });
-
-  // Restore from trash
-  ipcMain.handle('notes:restoreDeleted', async (_event, id: string) => {
-    const note = await repo.get(createNoteId(id));
-    if (!note) {
-      return { ok: false, error: { type: 'NOT_FOUND', id } };
-    }
-
-    const restoredNote = restoreDeletedNote(note);
-    await repo.save(restoredNote);
-
-    return { ok: true, data: noteToSnapshot(restoredNote) };
-  });
-
-  // Set note status
-  ipcMain.handle('notes:setStatus', async (_event, id: string, status: NoteStatus) => {
-    const note = await repo.get(createNoteId(id));
-    if (!note) {
-      return { ok: false, error: { type: 'NOT_FOUND', id } };
-    }
-
-    const updatedNote = setNoteStatus(note, status);
-    await repo.save(updatedNote);
-
-    return { ok: true, data: noteToSnapshot(updatedNote) };
-  });
-
-  // List notes
-  ipcMain.handle(
-    'notes:list',
-    async (
-      _event,
-      options?: {
-        limit?: number;
-        offset?: number;
-        tag?: string;
-        sortBy?: 'createdAt' | 'updatedAt' | 'title';
-        sortOrder?: 'asc' | 'desc';
-        archived?: 'active' | 'archived' | 'all';
-      }
-    ) => {
-      const notes = await repo.list(options);
-      return notes.map(note => noteToSnapshot(note));
-    }
-  );
-
-  // Search notes
-  ipcMain.handle('notes:search', async (_event, query: string, limit?: number) => {
-    const notes = await repo.search(query, limit);
-    return notes.map(note => noteToSnapshot(note));
-  });
-
-  // Get all tags
-  ipcMain.handle('notes:tags', async () => {
-    return repo.getAllTags();
-  });
-
-  // Set manual tags (full replacement)
-  ipcMain.handle('notes:setManualTags', async (_event, noteId: string, tags: string[]) => {
-    const id = createNoteId(noteId);
-    // Normalize tags: trim, lowercase, strip leading '#', remove empties, dedupe
-    const normalizedTags = [
-      ...new Set(tags.map(t => t.trim().toLowerCase().replace(/^#/, '')).filter(t => t.length > 0)),
-    ];
-    repo.setManualTags(
-      id,
-      normalizedTags.map(t => createTag(t))
-    );
-    return { ok: true };
-  });
-
-  // Get manual tags only (for editor to know which are removable)
-  ipcMain.handle('notes:getManualTags', async (_event, noteId: string) => {
-    const id = createNoteId(noteId);
-    return repo.getManualTags(id);
-  });
-
-  // Get all tags with colors
-  ipcMain.handle('tags:listWithColors', async () => {
-    return repo.getAllTagsWithColors();
-  });
-
-  // Set tag color
-  ipcMain.handle('tags:setColor', async (_event, tagName: string, color: string | null) => {
-    repo.setTagColor(tagName, color);
-    return { ok: true };
-  });
-
-  // Delete tag from system
-  ipcMain.handle('tags:delete', async (_event, tagName: string) => {
-    repo.deleteTag(tagName);
-    return { ok: true };
-  });
-
-  // Rename tag across all notes
-  ipcMain.handle('tags:rename', async (_event, oldName: string, newName: string) => {
-    return repo.renameTag(oldName, newName);
-  });
-
-  // ═══════════════════════════════════════════════════════════════════════════
-  // Links (Wikilinks / Backlinks)
-  // ═══════════════════════════════════════════════════════════════════════════
-
-  // Sync links for a note (call after saving note)
-  ipcMain.handle('links:sync', async (_event, noteId: string, content: string) => {
-    repo.syncLinks(createNoteId(noteId), content);
-    return { ok: true };
-  });
-
-  // Get backlinks (notes that link TO this note)
-  ipcMain.handle('links:backlinks', async (_event, noteId: string) => {
-    return repo.getBacklinks(createNoteId(noteId));
-  });
-
-  // Get outgoing links (notes this note links TO)
-  ipcMain.handle('links:outgoing', async (_event, noteId: string) => {
-    return repo.getOutgoingLinks(createNoteId(noteId));
-  });
-
-  // Get graph data (all notes and links for visualization)
-  ipcMain.handle('links:graph', async () => {
-    try {
-      return repo.getGraphData();
-    } catch (error) {
-      console.error('Failed to get graph data:', error);
-      // Return empty data on error
-      return { nodes: [], edges: [] };
-    }
-  });
-
-  // ═══════════════════════════════════════════════════════════════════════════
-  // Window Management
-  // ═══════════════════════════════════════════════════════════════════════════
-
+function registerWindowHandlers(): void {
   // Open a note in a new window
   ipcMain.handle('window:openNote', async (_event, noteId: string, noteTitle: string) => {
     createNoteWindow(noteId, noteTitle);
@@ -717,1947 +466,11 @@ function registerIpcHandlers(): void {
     createSettingsWindow();
     return { ok: true };
   });
-
-  // ═══════════════════════════════════════════════════════════════════════════
-  // Embeds (File Resolution)
-  // ═══════════════════════════════════════════════════════════════════════════
-
-  // Resolve embed target to asset:// URL
-  ipcMain.handle('embeds:resolve', async (_event, target: string, noteId: string) => {
-    if (!dataPaths) return null;
-
-    // Build path to note's assets folder: /assets/{noteId}/{target}
-    const assetPath = join(dataPaths.assets, noteId, target);
-
-    // Check if file exists
-    if (existsSync(assetPath)) {
-      // Return asset:// URL with host (required for browser to recognize protocol)
-      return `asset://local/${noteId}/${target}`;
-    }
-
-    // File not found
-    return null;
-  });
-
-  // Batch resolve multiple embed targets (more efficient)
-  ipcMain.handle(
-    'embeds:resolveBatch',
-    async (_event, targets: string[], noteId: string): Promise<Record<string, string | null>> => {
-      if (!dataPaths) return {};
-
-      const result: Record<string, string | null> = {};
-      for (const target of targets) {
-        const assetPath = join(dataPaths.assets, noteId, target);
-        // Return asset:// URL with host (required for browser to recognize protocol)
-        result[target] = existsSync(assetPath) ? `asset://local/${noteId}/${target}` : null;
-      }
-      return result;
-    }
-  );
-
-  // Save asset (image/file) for a note via drag & drop or paste
-  ipcMain.handle(
-    'embeds:saveAsset',
-    async (
-      _event,
-      noteId: string,
-      mime: string,
-      bytes: ArrayBuffer,
-      originalName?: string
-    ): Promise<{ ok: true; filename: string; relPath: string } | { ok: false; error: string }> => {
-      if (!dataPaths) {
-        return { ok: false, error: 'Data paths not initialized' };
-      }
-
-      // Validate noteId (non-empty, alphanumeric with hyphens/underscores)
-      if (!noteId || !/^[\w-]+$/.test(noteId)) {
-        return { ok: false, error: 'Invalid noteId' };
-      }
-
-      // Validate size (max 20MB)
-      const MAX_SIZE = 20 * 1024 * 1024;
-      if (bytes.byteLength > MAX_SIZE) {
-        return { ok: false, error: 'File too large (max 20MB)' };
-      }
-
-      // Derive extension from mime type
-      const mimeToExt: Record<string, string> = {
-        'image/png': 'png',
-        'image/jpeg': 'jpg',
-        'image/jpg': 'jpg',
-        'image/gif': 'gif',
-        'image/webp': 'webp',
-        'image/svg+xml': 'svg',
-        'image/bmp': 'bmp',
-        'image/ico': 'ico',
-        'video/mp4': 'mp4',
-        'video/webm': 'webm',
-        'video/quicktime': 'mov',
-        'audio/mpeg': 'mp3',
-        'audio/wav': 'wav',
-        'audio/ogg': 'ogg',
-        'application/pdf': 'pdf',
-      };
-
-      let ext = mimeToExt[mime];
-      if (!ext && originalName) {
-        // Fallback to originalName extension
-        const match = originalName.match(/\.([a-zA-Z0-9]+)$/);
-        ext = match?.[1]?.toLowerCase() ?? 'bin';
-      }
-      if (!ext) {
-        ext = 'bin';
-      }
-
-      // Generate unique filename: timestamp-random.ext
-      const timestamp = Date.now();
-      const random = Math.random().toString(36).substring(2, 8);
-      const filename = `${timestamp}-${random}.${ext}`;
-
-      // Ensure note's assets directory exists
-      const noteAssetsDir = join(dataPaths.assets, noteId);
-      await mkdir(noteAssetsDir, { recursive: true });
-
-      // Write file
-      const assetPath = join(noteAssetsDir, filename);
-      await writeFile(assetPath, Buffer.from(bytes));
-
-      return {
-        ok: true,
-        filename,
-        relPath: `${noteId}/${filename}`,
-      };
-    }
-  );
-
-  // Count notes
-  ipcMain.handle('notes:count', async () => {
-    // Get all notes to compute counts
-    const allNotes = await repo.list({ archived: 'all' });
-
-    const counts = {
-      active: 0,
-      archived: 0,
-      total: allNotes.length,
-      pinned: 0,
-      deleted: 0,
-      byStatus: {
-        active: 0,
-        on_hold: 0,
-        completed: 0,
-        dropped: 0,
-      } as Record<NoteStatus, number>,
-      byNotebook: {} as Record<string, number>,
-    };
-
-    for (const note of allNotes) {
-      // Count archived
-      if (note.metadata.archivedAt !== null) {
-        counts.archived++;
-      } else {
-        counts.active++;
-      }
-
-      // Count pinned
-      if (note.isPinned) {
-        counts.pinned++;
-      }
-
-      // Count deleted (in trash)
-      if (note.isDeleted) {
-        counts.deleted++;
-      }
-
-      // Count by status
-      if (note.status && counts.byStatus[note.status] !== undefined) {
-        counts.byStatus[note.status]++;
-      }
-
-      // Count by notebook (active, non-deleted notes only)
-      if (note.notebookId && !note.isDeleted && !note.metadata.archivedAt) {
-        counts.byNotebook[note.notebookId] = (counts.byNotebook[note.notebookId] || 0) + 1;
-      }
-    }
-
-    return counts;
-  });
 }
 
-/** Register IPC handlers for notebooks CRUD */
-function registerNotebookHandlers(): void {
-  if (!notebookRepository) {
-    throw new Error('Notebook repository not initialized');
-  }
-
-  const repo = notebookRepository;
-
-  // List all notebooks
-  ipcMain.handle('notebooks:list', async () => {
-    const notebooks = await repo.getAll();
-    return notebooks.map(nb => ({
-      id: nb.id,
-      name: nb.name,
-      parentId: nb.parentId,
-      depth: nb.depth,
-      order: nb.order,
-      createdAt: nb.createdAt,
-      updatedAt: nb.updatedAt,
-    }));
-  });
-
-  // Get notebook tree
-  ipcMain.handle('notebooks:tree', async () => {
-    return repo.getTree();
-  });
-
-  // Get single notebook
-  ipcMain.handle('notebooks:get', async (_event, id: string) => {
-    const notebook = await repo.get(createNotebookId(id));
-    if (!notebook) return null;
-    return {
-      id: notebook.id,
-      name: notebook.name,
-      parentId: notebook.parentId,
-      depth: notebook.depth,
-      order: notebook.order,
-      createdAt: notebook.createdAt,
-      updatedAt: notebook.updatedAt,
-    };
-  });
-
-  // Get notebook with metadata
-  ipcMain.handle('notebooks:getWithMetadata', async (_event, id: string) => {
-    const notebook = await repo.getWithMetadata(createNotebookId(id));
-    if (!notebook) return null;
-    return {
-      id: notebook.id,
-      name: notebook.name,
-      parentId: notebook.parentId,
-      depth: notebook.depth,
-      order: notebook.order,
-      createdAt: notebook.createdAt,
-      updatedAt: notebook.updatedAt,
-      noteCount: notebook.noteCount,
-      childCount: notebook.childCount,
-    };
-  });
-
-  // Create notebook
-  ipcMain.handle('notebooks:create', async (_event, input: { name: string; parentId?: string }) => {
-    let parentDepth = 0;
-    if (input.parentId) {
-      const parent = await repo.get(createNotebookId(input.parentId));
-      if (parent) {
-        parentDepth = parent.depth;
-      }
-    }
-
-    const nextOrder = await repo.getNextOrder(
-      input.parentId ? createNotebookId(input.parentId) : null
-    );
-
-    const notebook = createNotebook({
-      name: input.name,
-      parentId: input.parentId ? createNotebookId(input.parentId) : null,
-      parentDepth,
-      order: nextOrder,
-    });
-
-    await repo.save(notebook);
-
-    return {
-      id: notebook.id,
-      name: notebook.name,
-      parentId: notebook.parentId,
-      depth: notebook.depth,
-      order: notebook.order,
-      createdAt: notebook.createdAt,
-      updatedAt: notebook.updatedAt,
-    };
-  });
-
-  // Rename notebook
-  ipcMain.handle('notebooks:rename', async (_event, id: string, name: string) => {
-    const notebook = await repo.get(createNotebookId(id));
-    if (!notebook) {
-      throw new Error('Notebook not found');
-    }
-
-    const updated = renameNotebook(notebook, name);
-    await repo.save(updated);
-
-    return {
-      id: updated.id,
-      name: updated.name,
-      parentId: updated.parentId,
-      depth: updated.depth,
-      order: updated.order,
-      createdAt: updated.createdAt,
-      updatedAt: updated.updatedAt,
-    };
-  });
-
-  // Move notebook (recursively updates children's depth)
-  ipcMain.handle('notebooks:move', async (_event, id: string, newParentId: string | null) => {
-    const notebook = await repo.get(createNotebookId(id));
-    if (!notebook) {
-      throw new Error('Notebook not found');
-    }
-
-    // Prevent circular reference: can't move a notebook into its own descendant
-    if (newParentId) {
-      let current = await repo.get(createNotebookId(newParentId));
-      while (current && current.parentId) {
-        if (current.parentId === notebook.id) {
-          throw new Error('CIRCULAR_REFERENCE');
-        }
-        current = await repo.get(current.parentId);
-      }
-    }
-
-    let newParentDepth = 0;
-    if (newParentId) {
-      const parent = await repo.get(createNotebookId(newParentId));
-      if (parent) {
-        newParentDepth = parent.depth;
-      }
-    }
-
-    const result = moveNotebook(
-      notebook,
-      newParentId ? createNotebookId(newParentId) : null,
-      newParentDepth
-    );
-
-    if (!result.success) {
-      throw new Error(result.reason);
-    }
-
-    await repo.save(result.notebook);
-
-    // Recursively update children's depth to match the new hierarchy
-    const updateChildrenDepth = async (parentId: string, parentDepth: number) => {
-      const children = await repo.getChildren(parentId as ReturnType<typeof createNotebookId>);
-      for (const child of children) {
-        const newChildDepth = parentDepth + 1;
-        if (child.depth !== newChildDepth) {
-          await repo.save({ ...child, depth: newChildDepth });
-          await updateChildrenDepth(child.id, newChildDepth);
-        }
-      }
-    };
-    await updateChildrenDepth(result.notebook.id, result.notebook.depth);
-
-    return {
-      id: result.notebook.id,
-      name: result.notebook.name,
-      parentId: result.notebook.parentId,
-      depth: result.notebook.depth,
-      order: result.notebook.order,
-      createdAt: result.notebook.createdAt,
-      updatedAt: result.notebook.updatedAt,
-    };
-  });
-
-  // Delete notebook
-  ipcMain.handle('notebooks:delete', async (_event, id: string) => {
-    const notebookId = createNotebookId(id);
-
-    if (notebookId === INBOX_NOTEBOOK_ID) {
-      throw new Error('Cannot delete Inbox notebook');
-    }
-
-    await repo.delete(notebookId);
-    return { success: true };
-  });
-
-  // Reorder notebooks within a parent
-  ipcMain.handle(
-    'notebooks:reorder',
-    async (_event, parentId: string | null, orderedIds: string[]) => {
-      await repo.reorder(
-        parentId ? createNotebookId(parentId) : null,
-        orderedIds.map(id => createNotebookId(id))
-      );
-      return { success: true };
-    }
-  );
-
-  // ═══════════════════════════════════════════════════════════════════════════
-  // Git Operations
-  // ═══════════════════════════════════════════════════════════════════════════
-
-  // Enable git for a notebook
-  ipcMain.handle('notebooks:enableGit', async (_event, notebookId: string) => {
-    try {
-      repo.enableGit(createNotebookId(notebookId));
-      return { success: true };
-    } catch (error) {
-      return {
-        success: false,
-        error: error instanceof Error ? error.message : 'Failed to enable git',
-      };
-    }
-  });
-
-  // Disable git for a notebook
-  ipcMain.handle('notebooks:disableGit', async (_event, notebookId: string) => {
-    try {
-      repo.disableGit(createNotebookId(notebookId));
-      return { success: true };
-    } catch (error) {
-      return {
-        success: false,
-        error: error instanceof Error ? error.message : 'Failed to disable git',
-      };
-    }
-  });
-
-  // Check if git is enabled for a notebook
-  ipcMain.handle('notebooks:isGitEnabled', async (_event, notebookId: string) => {
-    try {
-      const enabled = repo.isGitEnabled(createNotebookId(notebookId));
-      return { success: true, enabled };
-    } catch (error) {
-      return {
-        success: false,
-        error: error instanceof Error ? error.message : 'Failed to check git status',
-      };
-    }
-  });
-
-  // Get git settings for a notebook
-  ipcMain.handle('notebooks:getGitSettings', async (_event, notebookId: string) => {
-    try {
-      const settings = repo.getGitSettings(createNotebookId(notebookId));
-      return { success: true, settings };
-    } catch (error) {
-      return {
-        success: false,
-        error: error instanceof Error ? error.message : 'Failed to get git settings',
-      };
-    }
-  });
-
-  // Toggle auto-commit for a notebook
-  ipcMain.handle(
-    'notebooks:setGitAutoCommit',
-    async (_event, notebookId: string, enabled: boolean) => {
-      try {
-        repo.setGitAutoCommit(createNotebookId(notebookId), enabled);
-        return { success: true };
-      } catch (error) {
-        return {
-          success: false,
-          error: error instanceof Error ? error.message : 'Failed to set auto-commit',
-        };
-      }
-    }
-  );
-
-  // Get all git-enabled notebooks
-  ipcMain.handle('notebooks:getGitEnabled', async () => {
-    try {
-      const notebooks = repo.getGitEnabledNotebooks();
-      return {
-        success: true,
-        notebooks: notebooks.map(nb => ({
-          id: nb.id,
-          name: nb.name,
-          parentId: nb.parentId,
-          depth: nb.depth,
-          order: nb.order,
-          createdAt: nb.createdAt,
-          updatedAt: nb.updatedAt,
-        })),
-      };
-    } catch (error) {
-      return {
-        success: false,
-        error: error instanceof Error ? error.message : 'Failed to get git-enabled notebooks',
-      };
-    }
-  });
-}
-
-/** Register IPC handlers for data management (backup, export, import) */
-function registerDataHandlers(): void {
-  if (!dataPaths || !noteRepository) {
-    throw new Error('Data paths or repository not initialized');
-  }
-
-  const paths = dataPaths;
-  const repo = noteRepository;
-
-  // Create backup
-  ipcMain.handle('data:backup', async () => {
-    return createBackup({
-      backupDir: paths.backups,
-      databasePath: paths.database,
-    });
-  });
-
-  // List backups
-  ipcMain.handle('data:backups:list', async () => {
-    return listBackups(paths.backups);
-  });
-
-  // Restore from backup
-  ipcMain.handle('data:backup:restore', async (_event, backupPath: string) => {
-    // Close current database connection
-    if (db) {
-      db.close();
-    }
-
-    const result = restoreBackup(backupPath, paths.database);
-
-    // Reconnect to database
-    db = createDatabase(paths.database);
-    runMigrations(db, allMigrations);
-
-    return result;
-  });
-
-  // Export notes
-  ipcMain.handle('data:export', async () => {
-    // Show save dialog
-    const { filePath, canceled } = await dialog.showSaveDialog({
-      title: 'Export Notes',
-      defaultPath: join(app.getPath('documents'), 'readied-export'),
-      buttonLabel: 'Export',
-    });
-
-    if (canceled || !filePath) {
-      return { success: false, error: 'Export cancelled' };
-    }
-
-    // Get all notes
-    const notes = await repo.list({ archived: 'all' });
-    const snapshots = notes.map(note => ({
-      id: note.id,
-      content: note.content,
-      title: note.title, // Use structural title
-      createdAt: note.metadata.createdAt,
-      updatedAt: note.metadata.updatedAt,
-      tags: [...note.metadata.tags],
-      wordCount: note.metadata.wordCount,
-      archivedAt: note.metadata.archivedAt,
-    }));
-
-    const result = exportNotes(snapshots, {
-      outputDir: filePath,
-      appVersion: app.getVersion(),
-      includeArchived: true,
-    });
-
-    if (result.success) {
-      // Open the export folder
-      shell.showItemInFolder(filePath);
-    }
-
-    return result;
-  });
-
-  // Export single note to file
-  ipcMain.handle(
-    'data:exportNote',
-    async (_event: Electron.IpcMainInvokeEvent, content: string, suggestedName: string) => {
-      let safeName =
-        suggestedName
-          .normalize('NFC')
-          // eslint-disable-next-line no-control-regex
-          .replace(/[/\\:*?"<>|\x00-\x1f.]/g, '')
-          .substring(0, 80)
-          .trim() || 'note';
-      const WINDOWS_RESERVED = /^(con|prn|aux|nul|com\d|lpt\d)$/i;
-      if (WINDOWS_RESERVED.test(safeName)) safeName = `_${safeName}`;
-      const { filePath, canceled } = await dialog.showSaveDialog({
-        title: 'Export Note',
-        defaultPath: join(app.getPath('documents'), `${safeName}.md`),
-        buttonLabel: 'Export',
-        filters: [{ name: 'Markdown', extensions: ['md'] }],
-      });
-
-      if (canceled || !filePath) {
-        return { success: false, error: 'Export cancelled' };
-      }
-
-      try {
-        await writeFile(filePath, content, 'utf-8');
-        return { success: true, path: filePath };
-      } catch (error) {
-        return {
-          success: false,
-          error: error instanceof Error ? error.message : 'Failed to write file',
-        };
-      }
-    }
-  );
-
-  // Import notes
-  ipcMain.handle('data:import', async () => {
-    // Show folder selection dialog
-    const { filePaths, canceled } = await dialog.showOpenDialog({
-      title: 'Import Notes',
-      properties: ['openDirectory'],
-      buttonLabel: 'Import',
-    });
-
-    const sourceDir = filePaths[0];
-    if (canceled || !sourceDir) {
-      return { success: false, error: 'Import cancelled' };
-    }
-
-    const importType = detectImportType(sourceDir);
-
-    const result = importNotes({
-      sourceDir,
-      type: importType,
-      recursive: true,
-    });
-
-    if (!result.success || !result.notes) {
-      return result;
-    }
-
-    // Import each note
-    let imported = 0;
-    for (const imported_note of result.notes) {
-      try {
-        await createNoteOperation(
-          {
-            content: imported_note.content,
-          },
-          repo
-        );
-        imported++;
-      } catch {
-        // Skip notes that fail to import
-      }
-    }
-
-    return {
-      success: true,
-      noteCount: imported,
-      skipped: result.skipped,
-    };
-  });
-
-  // Get data paths info
-  ipcMain.handle('data:paths', async () => {
-    return {
-      root: paths.root,
-      database: paths.database,
-      backups: paths.backups,
-      logs: paths.logs,
-    };
-  });
-
-  // Open data folder in system file manager
-  ipcMain.handle('data:openFolder', async () => {
-    void shell.openPath(paths.root);
-    return { success: true };
-  });
-}
-
-/** Register IPC handlers for renderer logging */
-function registerLogHandlers(): void {
-  const rendererLogger = createChildLogger({ component: 'renderer' });
-
-  // Log from renderer
-  ipcMain.handle(
-    'log:write',
-    async (
-      _event,
-      level: LogLevel,
-      message: string,
-      context?: Record<string, unknown>
-    ): Promise<{ success: boolean }> => {
-      const childLogger = context ? rendererLogger.child(context) : rendererLogger;
-
-      switch (level) {
-        case 'debug':
-          childLogger.debug(message);
-          break;
-        case 'info':
-          childLogger.info(message);
-          break;
-        case 'warn':
-          childLogger.warn(message);
-          break;
-        case 'error':
-          childLogger.error(message);
-          break;
-      }
-
-      return { success: true };
-    }
-  );
-
-  // Get log file path (for debugging/support)
-  ipcMain.handle('log:getPath', async (): Promise<string | null> => {
-    return dataPaths?.logs ?? null;
-  });
-}
-
-/** Register IPC handlers for manual update checks */
-function registerUpdateHandlers(): void {
-  // Manual check for updates
-  ipcMain.handle(
-    'updates:checkNow',
-    async (): Promise<{ available: boolean; version?: string }> => {
-      // In development or without proper updater config, return mock response
-      if (process.env.NODE_ENV === 'development') {
-        return { available: false };
-      }
-
-      // Event-based pattern: autoUpdater emits events, we wrap in Promise
-      return new Promise(resolve => {
-        const onAvailable = (info: { version: string }) => {
-          cleanup();
-          resolve({ available: true, version: info.version });
-        };
-
-        const onNotAvailable = () => {
-          cleanup();
-          resolve({ available: false });
-        };
-
-        const onError = () => {
-          cleanup();
-          resolve({ available: false });
-        };
-
-        const cleanup = () => {
-          autoUpdater.removeListener('update-available', onAvailable);
-          autoUpdater.removeListener('update-not-available', onNotAvailable);
-          autoUpdater.removeListener('error', onError);
-        };
-
-        autoUpdater.once('update-available', onAvailable);
-        autoUpdater.once('update-not-available', onNotAvailable);
-        autoUpdater.once('error', onError);
-
-        autoUpdater.checkForUpdates().catch(() => {
-          cleanup();
-          resolve({ available: false });
-        });
-      });
-    }
-  );
-
-  ipcMain.handle('updates:startDownload', async () => {
-    if (process.env.NODE_ENV === 'development') return { ok: false };
-    try {
-      await autoUpdater.downloadUpdate();
-      return { ok: true };
-    } catch (err) {
-      const message = (err as Error).message;
-      loggers.updater().error({ error: message }, 'Failed to download update');
-      return { ok: false, error: message };
-    }
-  });
-
-  ipcMain.handle('updates:installNow', () => {
-    // Force-close all windows so macOS doesn't block the quit
-    BrowserWindow.getAllWindows().forEach(win => {
-      if (!win.isDestroyed()) win.destroy();
-    });
-    autoUpdater.quitAndInstall(false, true);
-  });
-}
-
-/** Register IPC handlers for authentication and sync */
-function registerAuthSyncHandlers(): void {
-  if (!apiClient || !tokenStorage || !syncService) {
-    throw new Error('API client, token storage, or sync service not initialized');
-  }
-
-  const client = apiClient;
-  const storage = tokenStorage;
-  const sync = syncService;
-  const encryption = encryptionService;
-
-  // Broadcast sync status events to all renderer windows
-  sync.onStatusChange(event => {
-    broadcastToWindows('sync:status-changed', event);
-  });
-
-  // ═══════════════════════════════════════════════════════════════════════════
-  // Authentication
-  // ═══════════════════════════════════════════════════════════════════════════
-
-  // Request magic link email
-  ipcMain.handle('auth:requestMagicLink', async (_event, email: string) => {
-    try {
-      await client.requestMagicLink(email);
-      return { success: true };
-    } catch (error) {
-      return {
-        success: false,
-        error: error instanceof Error ? error.message : 'Failed to request magic link',
-      };
-    }
-  });
-
-  // Verify magic link token and save tokens
-  ipcMain.handle('auth:verify', async (_event, token: string) => {
-    try {
-      const result = await client.verifyMagicLink(token);
-      await storage.saveTokens(result.accessToken, result.refreshToken);
-      return { success: true, user: result.user };
-    } catch (error) {
-      return {
-        success: false,
-        error: error instanceof Error ? error.message : 'Failed to verify token',
-      };
-    }
-  });
-
-  // Get current session
-  ipcMain.handle('auth:getSession', async () => {
-    try {
-      const hasTokens = await storage.hasTokens();
-      if (!hasTokens) {
-        return null;
-      }
-
-      const user = await client.getCurrentUser();
-      return { user };
-    } catch (_error) {
-      // If session is invalid, clear tokens
-      await storage.clearTokens();
-      return null;
-    }
-  });
-
-  // Logout and clear tokens
-  ipcMain.handle('auth:logout', async () => {
-    try {
-      // Abort any in-flight sync operations before clearing tokens
-      sync?.stopAutoSync();
-      await storage.clearTokens();
-      return { success: true };
-    } catch (error) {
-      return {
-        success: false,
-        error: error instanceof Error ? error.message : 'Failed to logout',
-      };
-    }
-  });
-
-  // Refresh access token
-  ipcMain.handle('auth:refreshToken', async () => {
-    try {
-      const refreshed = await client.refreshAccessToken();
-      return { success: refreshed };
-    } catch (_error) {
-      return { success: false };
-    }
-  });
-
-  // ═══════════════════════════════════════════════════════════════════════════
-  // Sync
-  // ═══════════════════════════════════════════════════════════════════════════
-
-  // Pull changes from server
-  ipcMain.handle('sync:pull', async () => {
-    try {
-      const result = await sync.pull();
-      return {
-        success: result.success,
-        changes: result.changes,
-        cursor: result.cursor,
-        hasMore: result.hasMore,
-        conflicts: result.conflicts,
-        error: result.error,
-      };
-    } catch (error) {
-      return {
-        success: false,
-        error: error instanceof Error ? error.message : 'Failed to pull changes',
-      };
-    }
-  });
-
-  // Push changes to server
-  ipcMain.handle(
-    'sync:push',
-    async (
-      _event,
-      changes: Array<{
-        noteId: string;
-        operation: 'create' | 'update' | 'delete';
-        content?: string;
-        localVersion?: number;
-      }>
-    ) => {
-      try {
-        const result = await sync.push(changes);
-        return {
-          success: result.success,
-          results: result.results,
-          error: result.error,
-        };
-      } catch (error) {
-        return {
-          success: false,
-          error: error instanceof Error ? error.message : 'Failed to push changes',
-        };
-      }
-    }
-  );
-
-  // Perform full sync (pull + push)
-  ipcMain.handle('sync:syncNow', async () => {
-    try {
-      const result = await sync.syncNow();
-      return result;
-    } catch (error) {
-      return {
-        success: false,
-        changesApplied: 0,
-        changesPushed: 0,
-        conflicts: [],
-        error: error instanceof Error ? error.message : 'Sync failed',
-      };
-    }
-  });
-
-  // Get sync status
-  ipcMain.handle('sync:status', async () => {
-    try {
-      const state = sync.getState();
-      return {
-        success: true,
-        cursor: state.cursor,
-        lastSyncAt: state.lastSyncAt,
-        isSyncing: state.isSyncing,
-        lastError: state.lastError,
-        consecutiveFailures: state.consecutiveFailures,
-      };
-    } catch (error) {
-      return {
-        success: false,
-        error: error instanceof Error ? error.message : 'Failed to get sync status',
-      };
-    }
-  });
-
-  // Get pending change count (offline queue size)
-  ipcMain.handle('sync:pendingCount', async () => {
-    try {
-      return { success: true, count: sync.getPendingCount() };
-    } catch (error) {
-      return { success: false, count: 0, error: error instanceof Error ? error.message : 'Failed' };
-    }
-  });
-
-  // Resolve conflict
-  ipcMain.handle(
-    'sync:resolveConflict',
-    async (_event, noteId: string, resolution: 'local' | 'remote') => {
-      try {
-        await sync.resolveConflict(noteId, resolution);
-        return {
-          success: true,
-        };
-      } catch (error) {
-        return {
-          success: false,
-          error: error instanceof Error ? error.message : 'Failed to resolve conflict',
-        };
-      }
-    }
-  );
-
-  // Start auto-sync
-  ipcMain.handle('sync:startAutoSync', async (_event, intervalMs?: number) => {
-    try {
-      sync.startAutoSync(intervalMs);
-      return {
-        success: true,
-      };
-    } catch (error) {
-      return {
-        success: false,
-        error: error instanceof Error ? error.message : 'Failed to start auto-sync',
-      };
-    }
-  });
-
-  // Stop auto-sync
-  ipcMain.handle('sync:stopAutoSync', async () => {
-    try {
-      sync.stopAutoSync();
-      return {
-        success: true,
-      };
-    } catch (error) {
-      return {
-        success: false,
-        error: error instanceof Error ? error.message : 'Failed to stop auto-sync',
-      };
-    }
-  });
-
-  // Tag sync - pull
-  ipcMain.handle('sync:pullTags', async () => {
-    try {
-      return await sync.pullTags();
-    } catch (error) {
-      return { success: false, applied: 0, error: String(error) };
-    }
-  });
-
-  // Tag sync - push
-  ipcMain.handle('sync:pushTags', async () => {
-    try {
-      return await sync.pushTags();
-    } catch (error) {
-      return { success: false, pushed: 0, error: String(error) };
-    }
-  });
-
-  ipcMain.handle('sync:history', async (_event, limit?: number) => {
-    try {
-      const history = sync.getSyncHistory(limit);
-      return { success: true, history };
-    } catch (error) {
-      return {
-        success: false,
-        history: [],
-        error: error instanceof Error ? error.message : 'Failed to get sync history',
-      };
-    }
-  });
-
-  // ═══════════════════════════════════════════════════════════════════════════
-  // E2EE Key Management
-  // ═══════════════════════════════════════════════════════════════════════════
-
-  // Check if encryption is ready (CEK cached locally)
-  ipcMain.handle('encryption:isReady', async () => {
-    return { ready: encryption?.isReady() ?? false };
-  });
-
-  // Check if this is a first-time setup or existing user
-  ipcMain.handle('encryption:getKeyStatus', async () => {
-    try {
-      const serverKeys = await client.getEncryptionKeys();
-      const hasLocalKey = encryption?.isReady() ?? false;
-      const hasLegacyKey = encryption?.hasLegacyKey() ?? false;
-
-      return {
-        success: true,
-        hasServerKeys: serverKeys.exists,
-        hasLocalKey,
-        hasLegacyKey,
-      };
-    } catch (error) {
-      return {
-        success: false,
-        error: error instanceof Error ? error.message : 'Failed to get key status',
-      };
-    }
-  });
-
-  // First device: set up encryption keys with passphrase
-  ipcMain.handle('encryption:setupKeys', async (_event, passphrase: string) => {
-    try {
-      if (!encryption) throw new Error('Encryption service not available');
-
-      const result = await encryption.setupKeys(passphrase);
-
-      // Upload to server
-      await client.setEncryptionKeys({
-        salt: result.salt,
-        wrappedCek: result.wrappedCek,
-        wrappedCekRecovery: result.wrappedCekRecovery,
-        kdfParams: result.kdfParams,
-      });
-
-      return {
-        success: true,
-        recoveryKey: result.recoveryKey, // Show once to user!
-      };
-    } catch (error) {
-      return {
-        success: false,
-        error: error instanceof Error ? error.message : 'Failed to setup encryption keys',
-      };
-    }
-  });
-
-  // New device: unlock with passphrase
-  ipcMain.handle('encryption:unlockWithPassphrase', async (_event, passphrase: string) => {
-    try {
-      if (!encryption) throw new Error('Encryption service not available');
-
-      const serverKeys = await client.getEncryptionKeys();
-      if (
-        !serverKeys.exists ||
-        !serverKeys.salt ||
-        !serverKeys.wrappedCek ||
-        !serverKeys.kdfParams
-      ) {
-        return { success: false, error: 'No encryption keys found on server' };
-      }
-
-      await encryption.unlockWithPassphrase(
-        passphrase,
-        serverKeys.salt,
-        serverKeys.wrappedCek,
-        serverKeys.kdfParams
-      );
-
-      return { success: true };
-    } catch (error) {
-      const msg = error instanceof Error ? error.message : 'Failed to unlock';
-      const isWrongPassphrase = msg.includes('incorrect passphrase') || msg.includes('unwrap');
-      return {
-        success: false,
-        wrongPassphrase: isWrongPassphrase,
-        error: isWrongPassphrase ? 'Incorrect passphrase' : msg,
-      };
-    }
-  });
-
-  // Unlock with recovery key
-  ipcMain.handle('encryption:unlockWithRecoveryKey', async (_event, recoveryKey: string) => {
-    try {
-      if (!encryption) throw new Error('Encryption service not available');
-
-      const serverKeys = await client.getEncryptionKeys();
-      if (!serverKeys.exists || !serverKeys.wrappedCekRecovery) {
-        return { success: false, error: 'No recovery key found on server' };
-      }
-
-      await encryption.unlockWithRecoveryKey(recoveryKey, serverKeys.wrappedCekRecovery);
-
-      return { success: true };
-    } catch (error) {
-      return {
-        success: false,
-        error: error instanceof Error ? error.message : 'Failed to unlock with recovery key',
-      };
-    }
-  });
-
-  // Migrate legacy per-device key to key hierarchy
-  ipcMain.handle('encryption:migrateLegacyKey', async (_event, passphrase: string) => {
-    try {
-      if (!encryption) throw new Error('Encryption service not available');
-
-      const result = await encryption.migrateLegacyKey(passphrase);
-
-      // Upload to server
-      await client.setEncryptionKeys({
-        salt: result.salt,
-        wrappedCek: result.wrappedCek,
-        wrappedCekRecovery: result.wrappedCekRecovery,
-        kdfParams: result.kdfParams,
-      });
-
-      return {
-        success: true,
-        recoveryKey: result.recoveryKey,
-      };
-    } catch (error) {
-      return {
-        success: false,
-        error: error instanceof Error ? error.message : 'Failed to migrate legacy key',
-      };
-    }
-  });
-
-  // Change passphrase (re-wrap CEK)
-  ipcMain.handle('encryption:changePassphrase', async (_event, newPassphrase: string) => {
-    try {
-      if (!encryption) throw new Error('Encryption service not available');
-
-      const result = await encryption.changePassphrase(newPassphrase);
-
-      // Upload new wrapped key to server
-      await client.setEncryptionKeys({
-        salt: result.salt,
-        wrappedCek: result.wrappedCek,
-        kdfParams: result.kdfParams,
-      });
-
-      return { success: true };
-    } catch (error) {
-      return {
-        success: false,
-        error: error instanceof Error ? error.message : 'Failed to change passphrase',
-      };
-    }
-  });
-
-  // ═══════════════════════════════════════════════════════════════════════════
-  // Subscription
-  // ═══════════════════════════════════════════════════════════════════════════
-
-  // Get subscription status
-  ipcMain.handle('subscription:getStatus', async () => {
-    try {
-      const status = await client.getSubscriptionStatus();
-      return {
-        success: true,
-        status,
-      };
-    } catch (error) {
-      return {
-        success: false,
-        error: error instanceof Error ? error.message : 'Failed to get subscription status',
-      };
-    }
-  });
-
-  // Open Stripe billing portal
-  ipcMain.handle('subscription:openPortal', async (_event, returnUrl: string) => {
-    try {
-      const { url } = await client.createPortalSession(returnUrl);
-      void shell.openExternal(url);
-      return { success: true };
-    } catch (error) {
-      return {
-        success: false,
-        error: error instanceof Error ? error.message : 'Failed to open billing portal',
-      };
-    }
-  });
-
-  // Open checkout (placeholder - opens pricing page)
-  ipcMain.handle('subscription:openCheckout', async () => {
-    try {
-      void shell.openExternal('https://readied.app/pricing');
-      return { success: true };
-    } catch (error) {
-      return {
-        success: false,
-        error: error instanceof Error ? error.message : 'Failed to open checkout',
-      };
-    }
-  });
-
-  // ═══════════════════════════════════════════════════════════════════════════
-  // Devices
-  // ═══════════════════════════════════════════════════════════════════════════
-
-  ipcMain.handle('devices:list', async () => {
-    try {
-      const result = await client.listDevices();
-      return result.devices;
-    } catch (_error) {
-      return [];
-    }
-  });
-
-  ipcMain.handle('devices:rename', async (_event, deviceId: string, name: string) => {
-    try {
-      await client.renameDevice(deviceId, name);
-      return { success: true };
-    } catch (error) {
-      return {
-        success: false,
-        error: error instanceof Error ? error.message : 'Failed to rename device',
-      };
-    }
-  });
-
-  ipcMain.handle('devices:revoke', async (_event, deviceId: string) => {
-    try {
-      await client.revokeDevice(deviceId);
-      return { success: true };
-    } catch (error) {
-      return {
-        success: false,
-        error: error instanceof Error ? error.message : 'Failed to revoke device',
-      };
-    }
-  });
-
-  ipcMain.handle('devices:revokeOthers', async () => {
-    try {
-      const result = await client.revokeOtherDevices();
-      return { success: true, revokedCount: result.revokedCount };
-    } catch (error) {
-      return {
-        success: false,
-        error: error instanceof Error ? error.message : 'Failed to revoke devices',
-      };
-    }
-  });
-
-  ipcMain.handle('devices:getCurrent', async () => {
-    try {
-      const result = await client.listDevices();
-      return result.devices.find(d => d.isCurrent) ?? null;
-    } catch (_error) {
-      return null;
-    }
-  });
-
-  // ═══════════════════════════════════════════════════════════════════════════
-  // Encryption Key Management
-  // ═══════════════════════════════════════════════════════════════════════════
-
-  // Export encryption key (for backup)
-  ipcMain.handle('encryption:exportKey', async () => {
-    try {
-      if (!encryptionService) {
-        throw new Error('Encryption service not initialized');
-      }
-      const keyHex = encryptionService.exportKey();
-      return {
-        success: true,
-        key: keyHex,
-      };
-    } catch (error) {
-      return {
-        success: false,
-        error: error instanceof Error ? error.message : 'Failed to export encryption key',
-      };
-    }
-  });
-
-  // Import encryption key (for restore)
-  ipcMain.handle('encryption:importKey', async (_event, keyHex: string) => {
-    try {
-      if (!encryptionService) {
-        throw new Error('Encryption service not initialized');
-      }
-      await encryptionService.importKey(keyHex);
-      return {
-        success: true,
-      };
-    } catch (error) {
-      return {
-        success: false,
-        error: error instanceof Error ? error.message : 'Failed to import encryption key',
-      };
-    }
-  });
-}
-
-// ── AI Key Storage IPC Handlers ────────────────────────────────────────────
-
-ipcMain.handle('ai:saveKey', async (_event, provider: string, apiKey: string) => {
-  if (!aiKeyStorage) throw new Error('AI key storage not initialized');
-  await aiKeyStorage.saveKey(provider, apiKey);
-});
-
-ipcMain.handle('ai:getKey', async (_event, provider: string) => {
-  if (!aiKeyStorage) return null;
-  return aiKeyStorage.getKey(provider);
-});
-
-ipcMain.handle('ai:removeKey', async (_event, provider: string) => {
-  if (!aiKeyStorage) throw new Error('AI key storage not initialized');
-  await aiKeyStorage.removeKey(provider);
-});
-
-ipcMain.handle('ai:hasKey', async (_event, provider: string) => {
-  if (!aiKeyStorage) return false;
-  return aiKeyStorage.hasKey(provider);
-});
-
-ipcMain.handle('ai:listConnectedProviders', async () => {
-  if (!aiKeyStorage) return [];
-  return aiKeyStorage.listProviders();
-});
-
-/** Register IPC handlers for git operations */
-function registerGitHandlers(): void {
-  if (!gitService) {
-    throw new Error('Git service not initialized');
-  }
-
-  const git = gitService;
-
-  // Initialize git repository for a notebook
-  ipcMain.handle('git:init', async (_event, notebookId: string) => {
-    try {
-      const repoPath = await git.initRepository(notebookId);
-      return {
-        success: true,
-        repoPath,
-      };
-    } catch (error) {
-      return {
-        success: false,
-        error: error instanceof Error ? error.message : 'Failed to initialize git repository',
-      };
-    }
-  });
-
-  // Check if notebook has git repository
-  ipcMain.handle('git:isRepo', async (_event, notebookId: string) => {
-    try {
-      const isRepo = await git.isGitRepository(notebookId);
-      return { success: true, isRepo };
-    } catch (error) {
-      return {
-        success: false,
-        error: error instanceof Error ? error.message : 'Failed to check git repository',
-      };
-    }
-  });
-
-  // Commit changes
-  ipcMain.handle(
-    'git:commit',
-    async (_event, notebookId: string, message: string, files?: string[]) => {
-      try {
-        const sha = await git.commit(notebookId, message, files);
-        return {
-          success: true,
-          sha,
-        };
-      } catch (error) {
-        return {
-          success: false,
-          error: error instanceof Error ? error.message : 'Failed to commit changes',
-        };
-      }
-    }
-  );
-
-  // Get commit history
-  ipcMain.handle('git:log', async (_event, notebookId: string, limit?: number) => {
-    try {
-      const commits = await git.log(notebookId, limit);
-      return {
-        success: true,
-        commits,
-      };
-    } catch (error) {
-      return {
-        success: false,
-        error: error instanceof Error ? error.message : 'Failed to get commit history',
-      };
-    }
-  });
-
-  // Get repository status
-  ipcMain.handle('git:status', async (_event, notebookId: string) => {
-    try {
-      const status = await git.status(notebookId);
-      return {
-        success: true,
-        status,
-      };
-    } catch (error) {
-      return {
-        success: false,
-        error: error instanceof Error ? error.message : 'Failed to get repository status',
-      };
-    }
-  });
-
-  // Checkout (revert to) a specific commit
-  ipcMain.handle('git:checkout', async (_event, notebookId: string, commitSha: string) => {
-    try {
-      await git.checkout(notebookId, commitSha);
-      return { success: true };
-    } catch (error) {
-      return {
-        success: false,
-        error: error instanceof Error ? error.message : 'Failed to checkout commit',
-      };
-    }
-  });
-
-  // Write note file to git repository
-  ipcMain.handle(
-    'git:writeNote',
-    async (_event, notebookId: string, noteId: string, content: string) => {
-      try {
-        await git.writeNoteFile(notebookId, noteId, content);
-        return { success: true };
-      } catch (error) {
-        return {
-          success: false,
-          error: error instanceof Error ? error.message : 'Failed to write note file',
-        };
-      }
-    }
-  );
-
-  // Read note file from git repository
-  ipcMain.handle('git:readNote', async (_event, notebookId: string, noteId: string) => {
-    try {
-      const content = await git.readNoteFile(notebookId, noteId);
-      return {
-        success: true,
-        content,
-      };
-    } catch (error) {
-      return {
-        success: false,
-        error: error instanceof Error ? error.message : 'Failed to read note file',
-      };
-    }
-  });
-
-  // Delete note file from git repository
-  ipcMain.handle('git:deleteNote', async (_event, notebookId: string, noteId: string) => {
-    try {
-      await git.deleteNoteFile(notebookId, noteId);
-      return { success: true };
-    } catch (error) {
-      return {
-        success: false,
-        error: error instanceof Error ? error.message : 'Failed to delete note file',
-      };
-    }
-  });
-}
-
-/** Register IPC handlers for plugin config persistence */
-function registerPluginConfigHandlers(): void {
-  ipcMain.handle('pluginConfig:get', (_event, pluginId: string, key: string) => {
-    const row = db!
-      .prepare('SELECT value FROM plugin_config WHERE plugin_id = ? AND key = ?')
-      .get(pluginId, key) as { value: string } | undefined;
-    return row ? JSON.parse(row.value) : undefined;
-  });
-
-  ipcMain.handle('pluginConfig:set', (_event, pluginId: string, key: string, value: unknown) => {
-    db!
-      .prepare('INSERT OR REPLACE INTO plugin_config (plugin_id, key, value) VALUES (?, ?, ?)')
-      .run(pluginId, key, JSON.stringify(value));
-  });
-
-  ipcMain.handle('pluginConfig:getAll', (_event, pluginId: string) => {
-    const rows = db!
-      .prepare('SELECT key, value FROM plugin_config WHERE plugin_id = ?')
-      .all(pluginId) as Array<{ key: string; value: string }>;
-    const result: Record<string, unknown> = {};
-    for (const row of rows) {
-      result[row.key] = JSON.parse(row.value);
-    }
-    return result;
-  });
-
-  ipcMain.handle('pluginConfig:clear', (_event, pluginId: string) => {
-    db!.prepare('DELETE FROM plugin_config WHERE plugin_id = ?').run(pluginId);
-  });
-}
-
-/** Register IPC handlers for plugin discovery */
-function registerPluginDiscoveryHandlers(): void {
-  if (!dataPaths || !db) {
-    throw new Error('Data paths or database not initialized');
-  }
-
-  const paths = dataPaths;
-  const database = db;
-
-  // Scan filesystem for plugins
-  ipcMain.handle('plugins:scan', async () => {
-    return scanPlugins(paths.plugins);
-  });
-
-  // Check if a plugin is enabled (default: true if no row exists)
-  ipcMain.handle('plugins:isEnabled', (_event, pluginId: string) => {
-    const row = database
-      .prepare('SELECT enabled FROM plugin_registry WHERE plugin_id = ?')
-      .get(pluginId) as { enabled: number } | undefined;
-    return row ? row.enabled === 1 : true;
-  });
-
-  // Set plugin enabled/disabled state
-  ipcMain.handle('plugins:setEnabled', (_event, pluginId: string, enabled: boolean) => {
-    database
-      .prepare(
-        'INSERT INTO plugin_registry (plugin_id, enabled) VALUES (?, ?) ON CONFLICT(plugin_id) DO UPDATE SET enabled = ?'
-      )
-      .run(pluginId, enabled ? 1 : 0, enabled ? 1 : 0);
-  });
-
-  // List all plugin registry state
-  ipcMain.handle('plugins:listState', () => {
-    const rows = database.prepare('SELECT plugin_id, enabled FROM plugin_registry').all() as Array<{
-      plugin_id: string;
-      enabled: number;
-    }>;
-    return rows.map(row => ({
-      pluginId: row.plugin_id,
-      enabled: row.enabled === 1,
-    }));
-  });
-
-  // Read init.js user script (returns null if not found)
-  ipcMain.handle('plugins:readInitScript', async () => {
-    const initPath = join(paths.root, 'init.js');
-    try {
-      const code = await readFile(initPath, 'utf-8');
-      return code;
-    } catch {
-      return null;
-    }
-  });
-
-  // Install plugin from archive (.tar.gz or .zip)
-  ipcMain.handle('plugins:install', async () => {
-    const { filePaths, canceled } = await dialog.showOpenDialog({
-      title: 'Install Plugin',
-      properties: ['openFile'],
-      filters: [{ name: 'Plugin Archive', extensions: ['tar.gz', 'tgz', 'zip'] }],
-      buttonLabel: 'Install',
-    });
-
-    if (canceled || !filePaths[0]) {
-      return { success: false, error: 'Cancelled' };
-    }
-
-    const archivePath = filePaths[0];
-    const fileName = basename(archivePath).toLowerCase();
-
-    // Hoist tmpDir so it can be cleaned up in finally
-    let tmpDir: string | null = null;
-
-    try {
-      // Ensure plugins dir exists
-      await mkdir(paths.plugins, { recursive: true });
-
-      // Extract to a temp dir first, then move validated plugin folder
-      tmpDir = join(paths.plugins, `__installing_${Date.now()}`);
-      const extractDir = tmpDir;
-      await mkdir(extractDir, { recursive: true });
-
-      await new Promise<void>((resolve, reject) => {
-        const cb = (error: Error | null) => {
-          if (error) reject(error);
-          else resolve();
-        };
-        if (fileName.endsWith('.zip')) {
-          if (process.platform === 'win32') {
-            execFile(
-              'powershell',
-              [
-                '-NoProfile',
-                '-NonInteractive',
-                '-Command',
-                'Expand-Archive',
-                '-Force',
-                '-Path',
-                archivePath,
-                '-DestinationPath',
-                extractDir,
-              ],
-              cb
-            );
-          } else {
-            execFile('unzip', ['-o', archivePath, '-d', extractDir], cb);
-          }
-        } else {
-          execFile('tar', ['-xzf', archivePath, '-C', extractDir], cb);
-        }
-      });
-
-      // Find the manifest.json — could be at root or one level deep
-      const entries = await readdir(extractDir);
-      let pluginSourceDir = extractDir;
-
-      // If there's a single subdirectory, use that as the plugin root
-      if (entries.length === 1 && entries[0]) {
-        const candidatePath = join(extractDir, entries[0]);
-        const candidateStat = await stat(candidatePath);
-        if (candidateStat.isDirectory()) {
-          pluginSourceDir = candidatePath;
-        }
-      }
-
-      // Validate: must have manifest.json
-      const manifestPath = join(pluginSourceDir, 'manifest.json');
-      if (!existsSync(manifestPath)) {
-        return { success: false, error: 'No manifest.json found in archive' };
-      }
-
-      const manifestRaw = await readFile(manifestPath, 'utf-8');
-      const manifest = JSON.parse(manifestRaw);
-      if (!manifest.id || !manifest.name) {
-        return { success: false, error: 'Invalid manifest: missing id or name' };
-      }
-
-      // Validate plugin ID - only allow alphanumeric, hyphens, underscores
-      if (!/^[a-zA-Z0-9_-]+$/.test(manifest.id)) {
-        return {
-          success: false,
-          error: 'Invalid plugin ID: must be alphanumeric with hyphens/underscores only',
-        };
-      }
-
-      // Verify path doesn't escape plugins directory
-      const destDir = join(paths.plugins, manifest.id);
-      if (!normalize(destDir).startsWith(normalize(paths.plugins))) {
-        return { success: false, error: 'Invalid plugin ID: path traversal detected' };
-      }
-
-      // Move to final destination
-      if (existsSync(destDir)) {
-        await rm(destDir, { recursive: true, force: true });
-      }
-
-      await rename(pluginSourceDir, destDir);
-
-      return { success: true, pluginId: manifest.id, pluginName: manifest.name };
-    } catch (error) {
-      return { success: false, error: String(error) };
-    } finally {
-      if (tmpDir && existsSync(tmpDir)) {
-        await rm(tmpDir, { recursive: true, force: true }).catch(() => {});
-      }
-    }
-  });
-
-  // Install plugin from a remote URL (marketplace download)
-  ipcMain.handle('plugins:installFromUrl', async (_event, url: string, pluginSlug: string) => {
-    // Safety: only allow https URLs
-    if (!url.startsWith('https://')) {
-      return { success: false, error: 'Only HTTPS URLs are allowed' };
-    }
-
-    // Ensure plugins dir exists
-    await mkdir(paths.plugins, { recursive: true });
-
-    // Download to a temp file inside the plugins dir
-    const tmpDir = join(paths.plugins, `__downloading_${Date.now()}`);
-    await mkdir(tmpDir, { recursive: true });
-
-    try {
-      const response = await net.fetch(url);
-      if (!response.ok) {
-        return { success: false, error: `Download failed: HTTP ${response.status}` };
-      }
-
-      // Limit download size to 50 MB
-      const MAX_PLUGIN_SIZE = 50 * 1024 * 1024;
-      const contentLength = response.headers.get('content-length');
-      if (contentLength && parseInt(contentLength, 10) > MAX_PLUGIN_SIZE) {
-        return { success: false, error: 'Plugin archive exceeds maximum size of 50 MB' };
-      }
-
-      const buffer = Buffer.from(await response.arrayBuffer());
-      if (buffer.byteLength > MAX_PLUGIN_SIZE) {
-        return { success: false, error: 'Plugin archive exceeds maximum size of 50 MB' };
-      }
-
-      // Determine archive type from URL pathname
-      const urlPathname = new URL(url).pathname.toLowerCase();
-      const isZip = urlPathname.endsWith('.zip');
-      const archiveExt = isZip ? '.zip' : '.tar.gz';
-      const archivePath = join(tmpDir, `plugin${archiveExt}`);
-      await writeFile(archivePath, buffer);
-
-      // Extract to a staging dir
-      const stageDir = join(tmpDir, 'extracted');
-      await mkdir(stageDir, { recursive: true });
-
-      await new Promise<void>((resolve, reject) => {
-        const cb = (error: Error | null) => {
-          if (error) reject(error);
-          else resolve();
-        };
-        if (isZip) {
-          if (process.platform === 'win32') {
-            execFile(
-              'powershell',
-              [
-                '-NoProfile',
-                '-NonInteractive',
-                '-Command',
-                'Expand-Archive',
-                '-Force',
-                '-Path',
-                archivePath,
-                '-DestinationPath',
-                stageDir,
-              ],
-              cb
-            );
-          } else {
-            execFile('unzip', ['-o', archivePath, '-d', stageDir], cb);
-          }
-        } else {
-          execFile('tar', ['-xzf', archivePath, '-C', stageDir], cb);
-        }
-      });
-
-      // Find manifest.json — could be at root or one level deep
-      const entries = await readdir(stageDir);
-      let pluginSourceDir = stageDir;
-
-      if (entries.length === 1 && entries[0]) {
-        const candidatePath = join(stageDir, entries[0]);
-        const candidateStat = await stat(candidatePath);
-        if (candidateStat.isDirectory()) {
-          pluginSourceDir = candidatePath;
-        }
-      }
-
-      // Validate: must have manifest.json
-      const manifestPath = join(pluginSourceDir, 'manifest.json');
-      if (!existsSync(manifestPath)) {
-        return { success: false, error: 'No manifest.json found in downloaded archive' };
-      }
-
-      const manifestRaw = await readFile(manifestPath, 'utf-8');
-      const manifest = JSON.parse(manifestRaw);
-      if (typeof manifest !== 'object' || manifest === null || Array.isArray(manifest)) {
-        return { success: false, error: 'Invalid manifest: not a JSON object' };
-      }
-      if (!manifest.id || !manifest.name) {
-        return { success: false, error: 'Invalid manifest: missing id or name' };
-      }
-
-      // Cross-plugin overwrite protection: if we requested plugin A but the
-      // archive contains plugin B, block when it would overwrite an existing plugin
-      if (pluginSlug && manifest.id !== pluginSlug) {
-        const wouldOverwrite = join(paths.plugins, manifest.id);
-        if (existsSync(wouldOverwrite)) {
-          return {
-            success: false,
-            error: `Archive contains "${manifest.id}" but "${pluginSlug}" was requested. Refusing to overwrite existing plugin.`,
-          };
-        }
-      }
-
-      // Validate plugin ID - only allow alphanumeric, hyphens, underscores
-      if (!/^[a-zA-Z0-9_-]+$/.test(manifest.id)) {
-        return {
-          success: false,
-          error: 'Invalid plugin ID: must be alphanumeric with hyphens/underscores only',
-        };
-      }
-
-      // Verify path doesn't escape plugins directory
-      const destDir = join(paths.plugins, manifest.id);
-      if (!normalize(destDir).startsWith(normalize(paths.plugins))) {
-        return { success: false, error: 'Invalid plugin ID: path traversal detected' };
-      }
-
-      // Move to final destination
-      if (existsSync(destDir)) {
-        await rm(destDir, { recursive: true, force: true });
-      }
-
-      await rename(pluginSourceDir, destDir);
-
-      return {
-        success: true,
-        pluginId: manifest.id,
-        pluginName: manifest.name,
-        slugMismatch: pluginSlug && manifest.id !== pluginSlug ? pluginSlug : undefined,
-      };
-    } catch (error) {
-      return { success: false, error: String(error) };
-    } finally {
-      // Always clean up temp dir
-      if (existsSync(tmpDir)) {
-        await rm(tmpDir, { recursive: true, force: true }).catch(() => {});
-      }
-    }
-  });
-
-  // Uninstall plugin (remove its directory)
-  ipcMain.handle('plugins:uninstall', async (_event, pluginId: string) => {
-    // Safety: only allow removing from the plugins directory, prevent path traversal
-    const safeName = pluginId.replace(/[^a-z0-9-]/g, '');
-    const pluginDir = join(paths.plugins, safeName);
-    const normalizedDir = normalize(pluginDir);
-
-    if (!normalizedDir.startsWith(normalize(paths.plugins))) {
-      return { success: false, error: 'Invalid plugin ID' };
-    }
-
-    if (!existsSync(pluginDir)) {
-      return { success: false, error: 'Plugin not found' };
-    }
-
-    try {
-      await rm(pluginDir, { recursive: true, force: true });
-      // Clean up registry entry
-      database.prepare('DELETE FROM plugin_registry WHERE plugin_id = ?').run(pluginId);
-      return { success: true };
-    } catch (error) {
-      return { success: false, error: String(error) };
-    }
-  });
-
-  // Request plugin reload: broadcast to all windows except sender
-  ipcMain.on('plugins:requestReload', event => {
-    const senderWebContents = event.sender;
-    for (const win of BrowserWindow.getAllWindows()) {
-      try {
-        if (
-          win.webContents !== senderWebContents &&
-          !win.isDestroyed() &&
-          !win.webContents.isDestroyed()
-        ) {
-          win.webContents.send('plugins:reload');
-        }
-      } catch {
-        // Window destroyed during iteration
-      }
-    }
-  });
-}
-
-/** Initialize auto-updater */
-function initAutoUpdater(): void {
-  const updateLog = loggers.updater();
-
-  // Only check for updates in production
-  if (process.env.NODE_ENV === 'development') {
-    updateLog.debug('Skipping auto-updater in development');
-    return;
-  }
-
-  autoUpdater.autoDownload = false;
-  autoUpdater.autoInstallOnAppQuit = true;
-
-  autoUpdater.on('checking-for-update', () => {
-    updateLog.info('Checking for updates...');
-  });
-
-  autoUpdater.on('update-available', info => {
-    updateLog.info({ version: info.version }, 'Update available');
-    broadcastToWindows('updates:available', { version: info.version });
-  });
-
-  autoUpdater.on('update-not-available', () => {
-    updateLog.info('No updates available');
-  });
-
-  autoUpdater.on('download-progress', progress => {
-    updateLog.debug({ percent: progress.percent.toFixed(1) }, 'Download progress');
-    broadcastToWindows('updates:download-progress', {
-      percent: progress.percent,
-      bytesPerSecond: progress.bytesPerSecond,
-      transferred: progress.transferred,
-      total: progress.total,
-    });
-  });
-
-  autoUpdater.on('update-downloaded', info => {
-    updateLog.info({ version: info.version }, 'Update downloaded');
-    broadcastToWindows('updates:download-complete', { version: info.version });
-  });
-
-  autoUpdater.on('error', err => {
-    updateLog.error({ error: err.message }, 'Updater error');
-    broadcastToWindows('updates:error', { message: err.message });
-  });
-
-  // Check for updates after a short delay
-  setTimeout(() => {
-    void autoUpdater.checkForUpdates();
-  }, 3000);
-}
+// ============================================================================
+// Protocol Registration
+// ============================================================================
 
 /**
  * asset:// protocol
@@ -2690,7 +503,10 @@ protocol.registerSchemesAsPrivileged([
   },
 ]);
 
-// App lifecycle
+// ============================================================================
+// App Lifecycle
+// ============================================================================
+
 app
   .whenReady()
   .then(() => {
@@ -2726,13 +542,26 @@ app
     });
     log.info({ dataDir: dataPaths.root }, 'Application starting');
 
-    // Initialize database and handlers
+    // Initialize database and repositories
     initDatabase();
-    registerIpcHandlers();
-    registerNotebookHandlers();
-    registerGitHandlers(); // Git operations for git-backed notebooks
-    registerPluginConfigHandlers();
-    registerPluginDiscoveryHandlers();
+
+    // Register all IPC handler modules
+    registerNoteHandlers({
+      noteRepository: noteRepository!,
+      dataPaths,
+      noteToSnapshot,
+    });
+    registerNotebookHandlers({
+      notebookRepository: notebookRepository!,
+    });
+    registerGitHandlers({
+      gitService: gitService!,
+    });
+    registerPluginHandlers({
+      dataPaths,
+      db: db!,
+    });
+    registerWindowHandlers();
     registerAIHandlersNew(createAIService(), getToolRegistry());
 
     // Register built-in AI tools with database access
@@ -2775,12 +604,24 @@ app
       startPluginWatcher(dataPaths.plugins);
     }
 
-    registerDataHandlers();
+    registerDataHandlers({
+      dataPaths,
+      noteRepository: noteRepository!,
+      getDb: () => db!,
+      setDb: newDb => {
+        db = newDb;
+      },
+    });
 
     // Initialize license storage
     licenseStorage = new FileLicenseStorage(dataPaths.root);
-    registerLogHandlers();
-    registerUpdateHandlers();
+
+    registerLogHandlers({
+      getDataPaths: () => dataPaths,
+    });
+    registerUpdateHandlers({
+      broadcastToWindows,
+    });
 
     // App version
     ipcMain.handle('app:version', () => app.getVersion());
@@ -2858,8 +699,15 @@ app
           });
         }
 
-        registerAuthSyncHandlers();
+        registerAuthSyncHandlers({
+          apiClient,
+          tokenStorage,
+          syncService,
+          encryptionService,
+          broadcastToWindows,
+        });
         registerShareHandlers({ apiClient });
+        registerAiKeyHandlers({ aiKeyStorage });
         log.info('Auth and sync services initialized');
       } catch (error) {
         log.error(
@@ -2892,7 +740,7 @@ app
 
     // Create window and start auto-updater
     createWindow();
-    initAutoUpdater();
+    initAutoUpdater({ broadcastToWindows });
 
     app.on('activate', () => {
       if (BrowserWindow.getAllWindows().length === 0) {
